fix(hip-kernel-provider): address PR review feedback for runtime PBV scale
Simplify attn_scale_value fallback with value_or(), remove redundant
comment blocks, and change CreateNodeDirect compute type from BFLOAT16
to FLOAT in PBV scale test helpers for PR #10014 compatibility.

+++ dnn-providers/hip-kernel-provider/src/tests/engines/asm_sdpa_engine/TestSdpaBwdPlanBuilder.cpp
@@ -220,7 +220,7 @@ flatbuffers::FlatBufferBuilder createSdpaBwdGraphWithRuntimePbvScale()
     std::vector<flatbuffers::Offset<Node>> nodes;
     nodes.push_back(CreateNodeDirect(builder,
                                      "sdpa_bwd",
-                                     DataType::BFLOAT16,
+                                     DataType::FLOAT,
                                      NodeAttributes::SdpaBackwardAttributes,
                                      sdpaAttributes.Union()));
 
@@ -300,7 +300,7 @@ flatbuffers::FlatBufferBuilder createSdpaBwdGraphWithNonPbvScaleTensor()
     std::vector<flatbuffers::Offset<Node>> nodes;
     nodes.push_back(CreateNodeDirect(builder,
                                      "sdpa_bwd",
-                                     DataType::BFLOAT16,
+                                     DataType::FLOAT,
                                      NodeAttributes::SdpaBackwardAttributes,
                                      sdpaAttributes.Union()));
